pyproject.toml
[project]
name = "openproject-tui"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = ["textual (>=3.5.0,<4.0.0)", "httpx (>=0.28.1,<0.29.0)", "pydantic (>=2.11.7,<3.0.0)", "python-dotenv (>=1.1.1,<2.0.0)"]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
textual-dev = "^1.7.0"
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
pytest-mock = "^3.14.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
